+++ packages/project/src/merge/merge-project.ts
@@ -1,8 +1,9 @@
+import { randomUUID } from 'node:crypto';
 import { defaultsDeep, isEmpty } from 'lodash-es';
 import { CredentialState } from '@openfn/lexicon';
+import { Provisioner } from '@openfn/lexicon/lightning';
 import { Project } from '../Project';
 import { mergeWorkflows } from './merge-workflow';
-import { mergeCollections } from './merge-collections';
 import mapUuids from './map-uuids';
 import baseMerge from '../util/base-merge';
 import getDuplicates from '../util/get-duplicates';
@@ -184,6 +185,41 @@ export function merge(
   return new Project(
     baseMerge(target, source, ['channels'], assigns as any)
   );
+}
+
+/**
+ * Diff a locally-edited list of collection names (from openfn.yaml, no ids)
+ * against the collections that actually exist on the remote project (from a
+ * fresh fetch, with real ids), matched by name.
+ *
+ * A local name with no match remotely is a new collection (gets a fresh
+ * id). A remote collection with no matching local name is no longer wanted
+ * and is flagged for deletion. Renaming is not supported: it's seen as a
+ * delete of the old name plus a create of the new one.
+ */
+export function mergeCollections(
+  source: string[] = [],
+  target: Provisioner.Collection[] = []
+): Provisioner.Collection[] {
+  const targetByName = new Map(target.map((c) => [c.name, c]));
+  const result: Provisioner.Collection[] = [];
+
+  for (const name of source) {
+    const existing = targetByName.get(name);
+    if (existing) {
+      result.push({ id: existing.id, name });
+      targetByName.delete(name);
+    } else {
+      result.push({ id: randomUUID(), name });
+    }
+  }
+
+  // anything left in targetByName was removed from the local list
+  for (const remaining of Array.from(targetByName.values())) {
+    result.push({ id: remaining.id, name: remaining.name, delete: true });
+  }
+
+  return result;
 }
 
 export const replaceCredentials = (
